+++ Course5_Project/Exercise-master/src/main/java/org/upgrad/services/UserServiceImpl.java
@@ -59,21 +59,7 @@ public class UserServiceImpl implements UserService {
         return userRepository.findUserID(username);
     }
 
-    @Override
-    public void userProfileDetails(int user_id, String firstName, String lastName, String aboutMe, Date dob, String contactNumber, String country) {
-        user_profile.userProfileDetails(user_id,firstName,lastName,aboutMe,dob,contactNumber,country);
-    }
 
-    @Override
-    public Iterable<UserProfile> getUserProfile(int id) {
-       return user_profile.getUserProfile(id);
-
-    }
-
-    @Override
-    public void deleteUserProfileById(int Id) {
-        user_profile.deleteUserProfileById(Id);
-    }
 
     @Override
     public void deleteUserById(int id) {
@@ -89,5 +75,15 @@ public class UserServiceImpl implements UserService {
     public void addCategory(String title, String desc) {
         userRepository.addCategory(title,desc);
     }
-  
+
+    @Override
+    public String findUserPassword(String username) {
+        return null;
+    }
+
+    @Override
+    public String findUserRole(String username) {
+        return null;
+    }
+
 }
